--- app/src/main/java/com/nbc/curtaincall/data/model/SearchListModel.kt
@@ -1,25 +1,0 @@
-package com.nbc.curtaincall.data.model
-
-import com.tickaroo.tikxml.annotation.Element
-import com.tickaroo.tikxml.annotation.PropertyElement
-import com.tickaroo.tikxml.annotation.Xml
-
-@Xml(name = "dbs")
-data class SearchListModel(
-    @Element(name = "db")
-    val searchShowList: List<SearchItem>?
-)
-
-@Xml(name = "db")
-data class SearchItem(
-    @PropertyElement(name = "area") val area: String?,
-    @PropertyElement(name = "fcltynm") val fcltynm: String?,
-    @PropertyElement(name = "genrenm") val genrenm: String?,
-    @PropertyElement(name = "mt20id") val mt20id: String?,
-    @PropertyElement(name = "openrun") val openrun: String?,
-    @PropertyElement(name = "poster") val poster: String?,
-    @PropertyElement(name = "prfnm") val prfnm: String?,
-    @PropertyElement(name = "prfpdfrom") val prfpdfrom: String?,
-    @PropertyElement(name = "prfpdto") val prfpdto: String?,
-    @PropertyElement(name = "prfstate") val prfstate: String?
-)
