+++ app/src/main/java/com/nbc/curtaincall/fetch/domain/repository/FetchRepository.kt
@@ -1,0 +1,5 @@
+package com.nbc.curtaincall.fetch.domain.repository
+
+interface FetchRepository {
+
+}
