+++ app/src/main/java/com/nbc/curtaincall/fetch/network/retrofit/RetrofitClient.kt
@@ -1,6 +1,9 @@
-package com.nbc.curtaincall.data.api
+package com.nbc.curtaincall.fetch.network.retrofit
 
 import com.nbc.curtaincall.BuildConfig
+import com.nbc.curtaincall.BuildConfig.KOPIS_API_KEY
+import com.nbc.curtaincall.fetch.remote.FetchRemoteDatasource
+import com.nbc.curtaincall.search.SearchRemoteDatasource
 import com.tickaroo.tikxml.TikXml
 import com.tickaroo.tikxml.retrofit.TikXmlConverterFactory
 import okhttp3.Interceptor
@@ -9,11 +12,10 @@ import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Retrofit
 import java.util.concurrent.TimeUnit
 
-
-
 private const val BASE_URL = "http://kopis.or.kr/openApi/restful/"
 private const val KOPIS_API_KEY = BuildConfig.KOPIS_API_KEY
-object SearchRetrotifClient {
+
+object RetrofitClient {
     private val loggingInterceptor = HttpLoggingInterceptor().apply {
         level = HttpLoggingInterceptor.Level.BODY // 로깅 레벨 설정 (BASIC, HEADERS, BODY)
     }
@@ -23,7 +25,7 @@ object SearchRetrotifClient {
         val original = chain.request()
         val originalHttpUrl = original.url
         val url = originalHttpUrl.newBuilder()
-            .addQueryParameter("service", BuildConfig.KOPIS_API_KEY)  // Kopis api key 기본 추가
+            .addQueryParameter("service", KOPIS_API_KEY)  // Kopis api key 기본 추가
             .build()
         val requestBuilder = original.newBuilder().url(url)
         val request = requestBuilder.build()
@@ -39,7 +41,6 @@ object SearchRetrotifClient {
         .connectTimeout(15, TimeUnit.SECONDS)
         .build()
 
-
     // Retrofit 설정
     private val retrofit by lazy {
         val parser = TikXml.Builder().exceptionOnUnreadXml(false).build()
@@ -49,6 +50,6 @@ object SearchRetrotifClient {
             .addConverterFactory(TikXmlConverterFactory.create(parser))
             .build()
     }
-
-    val kopisApi: KopisSearchApiInterface by lazy { retrofit.create(KopisSearchApiInterface::class.java) }
+    val fetch: FetchRemoteDatasource by lazy { retrofit.create(FetchRemoteDatasource::class.java) }
+    val search: SearchRemoteDatasource by lazy { retrofit.create(SearchRemoteDatasource::class.java) }
 }
