+++ app/src/main/java/com/nbc/curtaincall/ui/home/HomeViewModel.kt
@@ -4,29 +4,27 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.nbc.curtaincall.data.model.Boxof
-import com.nbc.curtaincall.data.model.Db
-import com.nbc.curtaincall.data.repository.BoxOfficeRepository
-import com.nbc.curtaincall.data.repository.ShowListRepository
+import com.nbc.curtaincall.fetch.model.BoxofResponse
+import com.nbc.curtaincall.fetch.model.DbResponse
+import com.nbc.curtaincall.fetch.remote.FetchRemoteDatasource
 import com.nbc.curtaincall.util.Constants
 import kotlinx.coroutines.launch
 
 class HomeViewModel(
-    private val showListRepository: ShowListRepository,
-    private val boxOfficeRepository: BoxOfficeRepository
+    private val fetchRemoteRepository: FetchRemoteDatasource,
 ) : ViewModel() {
-    private val _showList = MutableLiveData<List<Db>>()
-    val showList: LiveData<List<Db>> get() = _showList
+    private val _showList = MutableLiveData<List<DbResponse>>()
+    val showList: LiveData<List<DbResponse>> get() = _showList
 
-    private val _topRank = MutableLiveData<List<Boxof>>()
-    val topRank: LiveData<List<Boxof>> get() = _topRank
-    private val _genre = MutableLiveData<List<Db>>()
-    val genre: LiveData<List<Db>> get() = _genre
+    private val _topRank = MutableLiveData<List<BoxofResponse>>()
+    val topRank: LiveData<List<BoxofResponse>> get() = _topRank
+    private val _genre = MutableLiveData<List<DbResponse>>()
+    val genre: LiveData<List<DbResponse>> get() = _genre
 
     fun fetchUpcomingList() {
         viewModelScope.launch {
             runCatching {
-                _showList.value = showListRepository.fetchShowList().showList
+                _showList.value = fetchRemoteRepository.fetchShowList().showList
             }.onFailure {
 
             }
@@ -36,7 +34,7 @@ class HomeViewModel(
     fun fetchTopRank() {
         viewModelScope.launch {
             runCatching {
-                _topRank.value = boxOfficeRepository.fetchTopRank().boxof
+                _topRank.value = fetchRemoteRepository.fetchTopRank().boxof
             }.onFailure {
 
             }
@@ -46,7 +44,7 @@ class HomeViewModel(
     fun fetchGenre(genre: Int) {
         viewModelScope.launch {
             runCatching {
-                _genre.value = showListRepository.fetchGenre(shcate = getGenreCode(genre)).showList
+                _genre.value = fetchRemoteRepository.fetchGenre(shcate = getGenreCode(genre)).showList
             }.onFailure {
             }
         }
