+++ app/src/main/java/com/nbc/curtaincall/ui/search/SearchViewModel.kt
@@ -4,8 +4,8 @@ import android.util.Log
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.nbc.curtaincall.data.api.SearchRetrotifClient
-import com.nbc.curtaincall.data.model.SearchItem
+import com.nbc.curtaincall.fetch.network.retrofit.RetrofitClient
+import com.nbc.curtaincall.search.model.SearchItem
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
@@ -26,6 +26,6 @@ class SearchViewModel : ViewModel() {
     }
 
     suspend fun getSearchResult(search: String) = withContext(Dispatchers.IO) {
-        SearchRetrotifClient.kopisApi.getSearchFilterShowList(shprfnm = search).searchShowList
+        RetrofitClient.search.getSearchFilterShowList(shprfnm = search).searchShowList
     }
 }
